requirements.txt
boto3
aws-cdk.core
aws-cdk.aws_lambda
aws_cdk.aws_ssm